feat: update project and examples
