+++ src/DEPLOY.TemporalTables.EF.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.12.35514.174
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DEPLOY.TemporalTables.Console", "DEPLOY.TemporalTables.EF\DEPLOY.TemporalTables.Console.csproj", "{06D7474A-ADC5-4C5D-A9AA-D5215950903F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Debug|x64.Build.0 = Debug|Any CPU
+		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Debug|x86.Build.0 = Debug|Any CPU
+		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Release|x64.ActiveCfg = Release|Any CPU
+		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Release|x64.Build.0 = Release|Any CPU
+		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Release|x86.ActiveCfg = Release|Any CPU
+		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {76557097-E3EF-415D-8865-6DE8396B404D}
+	EndGlobalSection
+EndGlobal
