+++ src/DEPLOY.TemporalTables.EF/appSettings.json
@@ -1,0 +1,5 @@
+{
+  "ConnectionStrings": {
+    "DEPLOYConnection": ""
+  }
+}
