--- src/PoC.TempTables.EF.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.12.35514.174
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PoC.TempTables.EF", "PoC.TempTables.EF\PoC.TempTables.EF.csproj", "{06D7474A-ADC5-4C5D-A9AA-D5215950903F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{06D7474A-ADC5-4C5D-A9AA-D5215950903F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
