--- src/PoC.TempTables.EF/appSettings.json
@@ -1,5 +1,0 @@
-{
-  "ConnectionStrings": {
-    "DefaultConnection": ""
-  }
-}
